refactor(registry): adds registry for managing endpoints
